requirements.txt
vk_api
discord.py
pyyaml